start of work to add analysis views
