+++ spackmon/apps/main/views/__init__.py
@@ -1,0 +1,8 @@
+from .main import index, builds_by_tag, builds_by_owner, build_detail
+from .specs import spec_diff, spec_detail
+from .analysis import (
+    view_analysis_diffs,
+    view_analysis_results,
+    stability_test_package,
+    package_matrix,
+)
